#76 removed toSolidityInput to use formatProofForVerifierContract
