+++ contracts/contracts/verifiers/CreamVerifier.sol
@@ -47,12 +47,12 @@ contract CreamVerifier {
 
     }
 
-    function _verifyProof(
+    function verifyProof(
         uint[2] memory a,
         uint[2][2] memory b,
         uint[2] memory c,
         uint[2] memory input
-    ) internal view returns (bool r) {
+    ) external view returns (bool r) {
 
         Proof memory proof;
         proof.A = Pairing.G1Point(a[0], a[1]);
@@ -93,14 +93,5 @@ contract CreamVerifier {
             proof.C,
             vk.delta2
         );
-    }
-
-    function verifyProof(
-		bytes calldata proof,
-		uint[2] calldata inputs
-	) external view returns (bool r) {
-        // solidity does not support decoding uint[2][2] yet
-        (uint[2] memory a, uint[2] memory b1, uint[2] memory b2, uint[2] memory c) = abi.decode(proof, (uint[2], uint[2], uint[2], uint[2]));
-        return _verifyProof(a, [b1, b2], c, inputs);
     }
 }
